refactor into a main group (#9)
