+++ app/(main)/not-found.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Link from "next/link";
-import { Section } from "./components/Section";
+import { Section } from "../components/Section";
 
 export default function NotFound() {
   return (
